fix: 피드백에 맞춰서 수정

+++ src/main/java/umc/domain/mission/repository/MemberMissionRepository.java
@@ -4,13 +4,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import umc.domain.member.entity.Member;
-import umc.domain.mission.entity.Mission;
 import umc.domain.mission.entity.mapping.MemberMission;
 import umc.domain.store.entity.Region;
 
 import java.util.List;
 
-public interface MemberMissionRepository extends JpaRepository<MemberMission,Integer> {
+public interface MemberMissionRepository extends JpaRepository<MemberMission,Long> {
     List<MemberMission> findAllByMemberAndIsCompleted(Member member, boolean isCompleted);
 
     @Query("SELECT COUNT(mm) FROM MemberMission mm WHERE mm.member.id=:memberId AND " +
